Settings: Fix Illegal State Exception in wifi advanced settings
Settings app crashes when the wifi advanced settings regulatory channels
list is opened and the device enters in to suspend mode and then the
settings process dies. This is because the list's
onRestoreInstanceState() tries to access the list array even before the
onResume() of wifi AdvancedSettings intializes it. Hence, initializing
this list in onCreate() to avoid this.

CRs-Fixed: 192208