Apps/Settings: Fix IndexOutOfBoundsException in ApnEditor
When a new ApnEditor window is opened and there is a MO/MT call, it
saves the instance of ApnEditor activity intent with action INSERT.
Next time if the same Apn is opened to edit and deleted it, onResume
of the intent with action INSERT is being called. When tried to save
its instance again, as the Apn is already deleted, the application
crashes. Hence destroying the activity if the save instance is called
on the intent with action INSERT.

CRs-Fixed: 212557