Settings: Ignore multiple service state change events in airplane mode
When airplane mode is turned on/off the settings app gets multiple
service state changed event causing the app to send multiple
RADIO_POWER requests to RIL without waiting for the response from
the previous request. This causes RIL to go in bad sate. The app now
ignores the service state change event that corresponds to the previous
state.

CRs-Fixed: 260338
Change-Id: I6748dae6d5c5790b6341df92809a5548f3c11e11