Bluetooth: Remember the state of 'Rename phone' activity on screen orientation change
Since we do not allow an empty name, when the Phone name is edited to zero
length, then the 'Rename' option in the activity will be grayed out. But,
when the orientation of the screen is changed with an empty phone name, the
option 'Rename' in the activity is enabled again. It is due the improper state
saved for the activity. Current change remembers the state properly and fixes
the issue.

Change-Id: I1b59c9f6026099420442e5acd3bd0bfefc602505
CRs-fixed: 341486