Bluetooth: Update Bluetooth device availability status message properly
In an out bound OPP transfer, when a device is selected from the device
picker through Gallery app, there is message "No nearby Bluetooth devices
were found" is shown though we had a list of device(s) and the device is
selected from the list. The change checks the status and updates the
message properly.

Change-Id: I61f171cea0eb129ba31ab2ff75bdd6e1bb2b8908
CRs-fixed: 342902