Settings(DSDS):Handle dialog of set data subscription
If user presses Home key when set data subscription dialog is
in progress then multisim settings activity goes to background
so dismisses the progress bar dialog. When set data subscription
is complete,this activity receives event and try to dismiss
progress bar dialog which is alredy dismissed resulting crash
in multisim settings.

Show or dismiss dialog accoding to activity in the foreground
or background

Change-Id: I239262733f57da45bdae9f112b5f7376848b9e4e
CRs-Fixed: 344429