Bluetooth : Remove out of range devices after ScanComplete
Change to remove device from UI when Bluetooth of remote
device is off or not discoverable anymore. Without this
change a device that discovered previously is available
in available devices list always, though the remote device
is not discoverable anymore. The current change updates
available devices list on scan complete removing the
devices that are available in cached devices and are not
disvorable anymore.

CRs-fixed: 342189

Change-Id: I37f67ad10da91b6a640e347b2f4f2edacb2e88ad