Settings(DSDS): Show error AlertDialog only activity in foreground
If user presses Home key when set data subscription dialog is
in progress then multisim settings activity goes to background.
When set data subscription is complete,this activity receives event
and try to show AlertDialog dialog which is resulting crash
in multisim settings.

Do not display the dialog boxes when the MultiSimSettings activity
is not in the foreground to avoid the runtime exception.

Change-Id: Ic259789d9ae35c2198cd559f384bc59aef39b128
CRs-Fixed: 366906