Bluetooth: Change MAP FTP CON REQ Checkbox String
Update MAP and FTP connection request check box string
from Don't ask again to Always Allow.
this change in string name is to ensure user does not have
wrong impression, because of similar strings displayed on
authorization request for different profiles but eventually
the behavior is different.
As per PBAP implementation once the connection REQ is
rejected by remote and checkbox is checked,
connection req should always be rejected till DUT is rebonded
with PBAP client and connection should always be accepted
if checkbox is checked and YES option is selected.
Whereas for QC implemented MAP and FTP profile this checkbox
will have effect only when its checked and YES button is
selected, which ensures connection would be auto accepted
from next time onwards and if the checkbox is checked but
NO button is pressed then connection would be rejected and
checkbox won't have any effect on the next connection request.
Hence change the text to appropriate one to be at par with
the actual implementation.

CRs-Fixed: 363888

Change-Id: I18cbac432b056dec1b0ec1e917eeef53af7574a0