Settings: set network checkbox status and do not popup dialog.
If enable 'access to my location', the google location will
popup a confirm alertdialog, and the app popups a dialog when
preference is changed, so it will popup up dialog twice. And it
doesn't set the network checkbox's status when the master toggle
is changed, so "WIFI & mobile network location" will be disabled.

Do not popup dialog in app, and set network checkbox's status
when the master toggle is changed.

CRs-Fixed: 497960

Change-Id: I8fad59bbddc30f85871c8f60745083844c12d610