Settings: Add multi-language APN name support
Add a new column "localized_name" to the carriers database for
multi-language APN name support.

The value of localized_name refers to a string resource id in
Strings.xml.

If a localized APN name is available, the dispalyed APN name
will be replaced to the localized APN name. A localized APN
name is unchangeable.

Change-Id: Ifa9aecaef97b2f800deb21ea9ee0a46d3c3d0910