Settings: Save a long word(>48 letters) to dictionary works well
The length of the word which is going to add to dictionary
is limited to a maximum of 48 characters. The set selection
action will throw a IndexOutOfBoundsException when the word
length is more than 48.

When the word length is longer than the maximum, move the cursor
to the end of actural word which is shown in edittext.

CRs-Fixed: 519526

Change-Id: I182a573b44484e10fb71a956204de45a7b1fbf36