Bluetooth: Remove out of range devices after ScanComplete
Change to remove device from UI when Bluetooth of remote
device is off or not discoverable anymore. Without this
change a device that discovered previously is available
in available devices list always, though the remote device
is not discoverable anymore. The current change updates
available devices list on scan complete removing the
devices that are available in cached devices and are not
disvorable anymore.

This reverts commit 1b652eb600ad7b4a769ad495bbb7eb4c6ab82fb2.

CRs-Fixed: 523374, 549186
Change-Id: Ia250d52e6fb7a69af242c961d33027f1e724ab25