Settings: Always displays "Visible to all nearby Bluetooth devices" if user choose to never timeout
As we set the alarm has not yet ended,during this time which we choose
to never timeout.When the time to alarm the scan mode will be set to
"SCAN_MODE_CONNECTABLE",so the message "Visible to all nearby Bluetooth
devices" will be changed.

Cancel alarm for BT discoverable when the visibility will be
setting never timeout.

CRs-Fixed: 536630

Change-Id: Ib243b2cec831a962c75edf995cb5f666622053c9