Settings: Fix screen color not changed after incoming calendar/alarm/call
When change screen color process is interrupted by incoming
calendar/alarm/call, the updated HSIC should be notified to
service. And when user press the back key, the updated HSIC
should be restored.

Do not restore the updated HSIC when incoming calendar/alarm/call.
And restore the updated HSIC when user press the back key.

CRs-Fixed: 541794

Change-Id: I07db115a9d621c594cd06898010091bcca73a49b