Bluetooth: Fixed the BT received pairing request should send a notification when the screen is OFF
When the BT received pairing request,it decided to send a notification
or pop-up a dialog according to screen is OFF.But this time the
"isScreenOn" method return value has been set to ON,so pop-up a dialog
prompts user does not have sound.

If the screen is OFF,then it is surely the lock screen state,we use the
lock screen to determine how to prompt the user.

CRs-Fixed: 508829

Change-Id: I35ca43ddb095f02d01f239930d1baca36ed1b390