Settings: Show "Operator info" for Brazil only
As "Operator info" is only used for Brazil to update latest area info,
only show area info when SIM country is Brazil.

Add a boolean switch to control whether latest area info is needed to
show. Check all the SIM card. If there is at least one SIM of Brazil,
set the switch to true to show the "Operator info". Otherwise, set the
switch to false to remove it from the preference list of phone status.

CRs-Fixed: 549188

Change-Id: Ifed912a2b913d65b8388c9840252c82edb4ecdc0