Bluetooth: disable scan icon when BT state is off
Usually onScanningStateChanged function should take care
of disabling the scan icon when the scan is completed.

But When the BT scan is going on and try to switch of the
BT, as onScanningStateChanged function doesn't have the
reference for scan icon it will not disable it, so disabling
is taken care in BluetoothSettings on receiving BT Off state.

CRs-Fixed: 543777
Change-Id: Ib827903d4f5c80eeb6ac8a06fb1a945140dd20c9