Settings: Fix the apn editor activity will crash when resume.
The validateAndSave will delete the data in cursor. When resume,
the activity will call fillUI to update the state of items. But
there is no data in cursor which causes the application crash.
Besides, the receiver doesn't unregister will also cause the
application crash.

The original design is only to update the state of item. Do not
call fillUI method and only update the state of list items.
Unregister the receiver to ensure the application works well.

CRs-Fixed: 537303

Change-Id: If9e15c37d780f61f59bcdccfb85845b6c508ec69