Settings: Fixed the BT tethering option can be enabled when locking screen
When we receive the BT state changed broadcast if the state is ON, we
will set up the BT tethering as checked. If we haven't received this
broadcast, then unlock screen the BT tethering will be unchecked.

When unlock screen to update the UI state, according to
"mBluetoothEnableForTether" to set up the BT tethering is available.

CRs-Fixed: 540044

Change-Id: I5c272c8b19357e81da82b374aa723772b9222ba8