Settings(MSIM): Update signal strength with correct dbm and asu values.
In the current implementation, when phone is in CDMA phone, EVDO dbm
value is not considered and when phone is in GSM mode, Lte and
TdsCdma values are not considered.

Also, in both modes, only GSM signal strength is considered for asu
value.

This fix updates signal strength with correct dbm and asu values.

Change-Id: I715e381e031cd7d7d5cc412e99e5e4b5ff3bf958
CRs-Fixed: 562772