Settings: Reset to default ringtone when data deleted.
The ringtone relies on the audio file on SD card when setting
audio on SD card as ringtone. When the SD card is unmount, we
can't query the data because the source file no longer exists.
In this situation, the summary string is set to unknown.

When the file is not exist, reset the ringtone to default one.

CRs-Fixed: 532104

Change-Id: Ida943d775b75e52d5f288f5e90f9bfff85c68dfa