Settings: Fix Settings force close on providing alphabetic values
Enter alphabetic text in editBox of the dialog when tap enableNetwork
causes Settings force close

The text of inputting in editBox will be converted to integer
variable. And the representation of integer variable
is -2147483648~2147483647. So if the text contains alphabetic values
or is larger than 2147483647 will leads to Settings force close. Add
protection to make sure the text of inputting is right.

CRs-fixed: 571836

Change-Id: If58c47a464d6157379142754b6264243519cd924