Settings: Fix "APN PPP phone number"is added after rotate device.
The source code will call the onCreate() method again in
ApnEditor.java when rotate device to landscape mode, and then the
fillUi() method will be called again, however, the "mFirstTime" is
change to false, can't remove the "APN PPP phone number" option.

Add the "configChanges" in AndroidManifest.xml to avoid to call
the onCreate() method again in ApnEditor.java when rotate device
to landscape mode.

CRs-fixed: 534704

Change-Id: Ie9dabb227dcaaf48ee62c333a06f054858c42793