Settings: Fix the bottom of interface will be in a mess issue.
When "Show owner info on lock screen" is disenabled, EditText get
focus and show inputmethods, and this will cause the bottom of
interface will be in a mess.

Fix the codes to judge the "Show owner info on lock screen"
checkbox whether be enabled, When "Show owner info on lockscreen"
is disenabled, EditText should not request focus and not show
inputmethods.

CRs-Fixed: 530942

Change-Id: I960ee3fe91bfc0ea7fdd66ef977f709e8ef568ee