Settings: No sound when change volume after check other sound
Because when the notificatioSettings activity is in background,
will call the onActivityStop() method to quit the current
looper, however, the SeekBarVolumizer instance doesn't set null,
so change volume after back from other activity, the handle
initialization can't be executed, throw IllegalStateException
which is sending message to a Handler on a dead thread.

Set SeekBarVolumizer instance null after call onActivityStop()
method so that handle initialization can be executed.

CRs-Fixed: 757850

Change-Id: I2bdf92926bc5c15d5b120b5ae2ab9c290bcacfaf