SimSettings: Cellular Data Preference
DDS switch triggered in an active CS session will fail PS attach
on the sub to switch to. Fixing this by disabling cellular
data prefrence when call state is not idle. A message is
displayed to the user indicating that cellular data preference is
disabled during an active voice call.

CRs-Fixed: 757906

Change-Id: I0a771e65d1e96835996b7f97ead37f4e864872eb