Settings: Fix cannot search paired device
Pair a device,the information of the device is inserted
into search_index.db.Then BluetoothEnabler.java invoke
updateFromClassNameResource() to update database.If
Bluetooth is on,the value of enable is true.If Bluetooth
is off,the value of enable is false.The value of enable
should be based on the status of the Bluetooth.

Delete excess code.

CRs-Fixed: 770172

Change-Id: I36d18e4f6c3a39529451874c40ffe25e4369f446