Wifi: Fetch the RSSI of the AP through getRSSI only in connected state
Framework shall start fetching the RSSI of the connected AP through the
signal_poll command only after the authentication is successful. Thus,
relying on this information before the authentication (during the
authenticating state ) shall lead to an improper RSSI value. Thus, get
RSSI from the WiFiInfo only when the station is in the connected state,
rather rely on the RSSI calculated through the scan results till then.

Change-Id: I2917ab96ad9a0d40055064f5582d46b02853c57e
CRs-Fixed: 781966