Add paired BT device name into Index database immediately
After we add BT device name to indexable data, it will only update the
Index database after we delete the original data. This is why we can not
search paired BT device name after we add it in to Indexable data.

Delete the database data before we add the paired BT device name into it.

Change-Id: Ibbaeb056db31eba84527e2677f7784195ac5921a
CRs-Fixed: 770172