Settings: Fix IllegalStateException in bluetooth device settings
The cause is that the size of mPendingActions in FragmentManager is
not 0. Beacause the DeviceProfilesSettings fragment is added to
mPendingActions when the function onCreate() is called. Then the
IllegalStateException(Recursive entry to executePendingTransactions)
will be thrown in function execPendingActions() when finish() is
called in DeviceProfilesSettings.

Replace the function finish() with getActivity().finish() in
onCreate(). Then mPendingActions will be 0.

Change-Id: Id1a05120389952ad29b729c0378bee5af9225464
CRs-Fixed: 795045