Settings: Hide screen color from DisplaySettings based on system property
1. Screen color is no more used for new hardware but it gives
conflicts to new APIs for HSIC settings. So hiding it for user
perspective based on system properties "persist.tuning.qdcm" and
"ro.qcom.screencolor".

2. If feature property ro.qcom.screencolor = 1, then only screen
color will be functional on the target.

3. Toggle property perist.tuning.qdcm is also there for those
hardware which have other tuning APIs for HSIC settings which can
give conflicts with screen color. If it is set to 1 then screen
color won't appear in device settings. Otherwise, it will appear
and work perfectly. It will also block other APIs to work.

CRs-Fixed: 804133
Change-Id: I3fd7e7afef991398d2b87d61c26110e2745c7af0