Settings: Fix FC when there isn't a string array translation.
The cause of this CR is that there is no translation for
string-array "screen_timeout_entries" in some language.The string
"Never" in the string-array should be translated in every language,
otherwise ArrayIndexOutOfBoundsException exists when the
screen timeout is set to never.

Override the function getEntry of ListPreference, make it return
string screensaver_timeout_zero_summary when the array index is
out of bounds.

Change-Id: Iff680d1e836bf50b856941ced2eae42d998f5a95
CRs-Fixed: 806722