Fix for Back key from Recent Location Apps returning to Launcher.
On Launching recent Location apps and pressing back key, we are
returning to launcher instead of going back to settings screen.
The FLAG_ACTIVITY_CLEAR_TASK is set in the intent when launching
the RecentLocationApps which is resulting in clearing of previous
task activities. When we use a different function in settings
app this flag is not included and issue is resolved.

CRs-Fixed: 893962
Change-Id: I955f4ef8e3b601417b7fd1bdac065aaaf8008d59